--- a/src/main/resources/reference.conf
+++ b/src/main/resources/reference.conf
@@ -36,6 +36,15 @@ jdbc-journal {
 
   dao = "akka.persistence.jdbc.journal.dao.ByteArrayJournalDao"
 
+  // The size of the buffer used when queueing up events for batch writing. This number must be bigger then the number
+  // of events that may be written concurrently. In other words this number must be bigger than the number of persistent
+  // actors that are actively peristing at the same time.
+  bufferSize = 1000
+  // The maximum size of the batches in which journal rows will be inserted
+  batchSize = 400
+  // The maximum number of batch-inserts that may be running concurrently
+  parallelism = 8
+
   slick {
     driver = "slick.jdbc.PostgresProfile$"
     db {
@@ -146,11 +155,6 @@ jdbc-read-journal {
   # How many events to fetch in one query (replay) and keep buffered until they
   # are delivered downstreams.
   max-buffer-size = "500"
-
-  # How many events should be written in the bulk insert
-  # 250 seems a sweet spot for a macbook pro i7 16GB that runs postgres
-  # on docker-platform for mac 1.12.0-rc4-beta20 (build: 10404)
-  batch-size = "250"
 
   dao = "akka.persistence.jdbc.query.dao.ByteArrayReadJournalDao"