--- a/src/test/resources/postgres-application.conf
+++ b/src/test/resources/postgres-application.conf
@@ -52,8 +52,6 @@ jdbc-read-journal {
 
   max-buffer-size = "500"
 
-  batch-size = "250"
-
   slick = ${slick}
   slick.db.numThreads = 20
   slick.db.maxConnections = 100