Implemented batch writing over multiple persistent actors. This results in a significant performance improvement.

I added result of the JDBCJournalPerfSpec below, I also added two testcases which test persist and persistAsync using a configurable amount of actors which are concurrently writing messages.

I would actually have wanted to test with more than 100 events, but the old implementation would quickly get circuit breaker timeouts...

This pull request also includes a second change: namely that it adheres to the spec for asyncReadHighestSequenceNr, which now reads the sequence number only if the previous write has completed (similar to akka-persistence-cassandra)

```
new implementation performance test output:
 PersistAsync()-ing 100 took 101 ms
 PersistAsync()-ing 100 took 37 ms
 PersistAsync()-ing 100 took 33 ms
 PersistAsync()-ing 100 took 36 ms
 PersistAsync()-ing 100 took 32 ms
 Average time: 48 ms
 Persist()-ing 100 took 1262 ms
 Persist()-ing 100 took 1324 ms
 Persist()-ing 100 took 1239 ms
 Persist()-ing 100 took 1263 ms
 Persist()-ing 100 took 1271 ms
 Average time: 1272 ms
 Recovering 100 took 23 ms
 Recovering 100 took 16 ms
 Recovering 100 took 16 ms
 Recovering 100 took 15 ms
 Recovering 100 took 15 ms
 Average time: 17 ms
 Persist()-ing 100 * 100 took 3372 ms
 Persist()-ing 100 * 100 took 3277 ms
 Persist()-ing 100 * 100 took 3211 ms
 Persist()-ing 100 * 100 took 3137 ms
 Persist()-ing 100 * 100 took 3088 ms
 Average time: 3217 ms
 persistAsync()-ing 100 * 100 took 347 ms
 persistAsync()-ing 100 * 100 took 351 ms
 persistAsync()-ing 100 * 100 took 287 ms
 persistAsync()-ing 100 * 100 took 300 ms
 persistAsync()-ing 100 * 100 took 317 ms
 Average time: 320 ms

old implementation performance test output
 PersistAsync()-ing 100 took 1401 ms
 PersistAsync()-ing 100 took 1271 ms
 PersistAsync()-ing 100 took 1294 ms
 PersistAsync()-ing 100 took 1285 ms
 PersistAsync()-ing 100 took 1263 ms
 Average time: 1303 ms
 Persist()-ing 100 took 1376 ms
 Persist()-ing 100 took 1344 ms
 Persist()-ing 100 took 1280 ms
 Persist()-ing 100 took 1254 ms
 Persist()-ing 100 took 1268 ms
 Average time: 1304 ms
 Recovering 100 took 18 ms
 Recovering 100 took 15 ms
 Recovering 100 took 14 ms
 Recovering 100 took 15 ms
 Recovering 100 took 13 ms
 Average time: 15 ms
 Persist()-ing 100 * 100 took 11020 ms
 Persist()-ing 100 * 100 took 11044 ms
 Persist()-ing 100 * 100 took 11068 ms
 Persist()-ing 100 * 100 took 10916 ms
 Persist()-ing 100 * 100 took 10936 ms
 Average time: 10997 ms
 persistAsync()-ing 100 * 100 took 11066 ms
 persistAsync()-ing 100 * 100 took 10831 ms
 persistAsync()-ing 100 * 100 took 11051 ms
 persistAsync()-ing 100 * 100 took 10846 ms
 persistAsync()-ing 100 * 100 took 10966 ms
 Average time: 10952 ms
```